Add: 「中を見てみる」ボタンの設置

+++ app/Http/Controllers/Auth/WithdrawController.php
@@ -17,22 +17,25 @@ class WithdrawController extends Controller
      */
     public function __invoke(Request $request)
     {
-        // Google APIを格納しているテーブルの不要な書籍を削除
-        $books =    DB::table('books as b')
-            ->select('b.id as book_id', 's.id as bookshelf_id')
-            ->leftJoin('bookshelf as s', 'b.id', '=', 's.book_id')
-            ->get();
 
-        foreach ($books as $key => $book) {
-            if ($book->bookshelf_id === null) {
-                DB::table('books')->where('id', $book->book_id)->delete();
+        if (Auth::id() !== 1) {
+            // Google APIを格納しているテーブルの不要な書籍を削除
+            $books =    DB::table('books as b')
+                ->select('b.id as book_id', 's.id as bookshelf_id')
+                ->leftJoin('bookshelf as s', 'b.id', '=', 's.book_id')
+                ->get();
+
+            foreach ($books as $key => $book) {
+                if ($book->bookshelf_id === null) {
+                    DB::table('books')->where('id', $book->book_id)->delete();
+                }
             }
+
+            // 退会処理
+            DB::table('users')->where('id', Auth::id())->delete();
+            DB::table('bookshelf')->where('user_id', Auth::id())->delete();
         }
 
-        // 退会処理
-        DB::table('users')->where('id', Auth::id())->delete();
-        DB::table('bookshelf')->where('user_id', Auth::id())->delete();
-
-        return redirect('login');
+        return redirect('edit/profile');
     }
 }
